fix: resolve CS0618 obsolete API warnings and CS0103/CS0169 errors
- Replace ElementId.IntegerValue with .Value across 7 files
- Cast BuiltInCategory to long for new ElementId() constructor (17 files)
- Fix StingListPicker.Show overload in LinkManagerCommands.cs (CS0103)
- Suppress CS0169 for unused fields in LANCollaborationCommands.cs
- Replace int.TryParse with long.TryParse for ElementId in BIMManagerCommands

+++ StingTools/BIMManager/BIMManagerCommands.cs
@@ -3850,7 +3850,7 @@ namespace StingTools.BIMManager
                         var elIds = r.RawJson["element_ids"] as JArray;
                         if (elIds != null)
                             foreach (var eid in elIds)
-                                if (int.TryParse(eid.ToString(), out int id))
+                                if (long.TryParse(eid.ToString(), out long id))
                                     ids.Add(new ElementId(id));
                     }
                     if (ids.Count > 0)
